src/test/java/com/davidbracewell/apollo/distribution/BinomialTest.java
/*
 * (c) 2005 David B. Bracewell
 *
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements.  See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership.  The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License.  You may obtain a copy of the License at
 *
 *   http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied.  See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */

package com.davidbracewell.apollo.distribution;

import org.junit.Before;
import org.junit.Test;

import static org.junit.Assert.*;

/**
 * @author David B. Bracewell
 */
public class BinomialTest {

   Binomial binomial;


   @Before
   public void setUp() throws Exception {
      binomial = new Binomial(5, 10);
   }


   @Test
   public void probability() throws Exception {
      assertEquals(0.205, binomial.probability(6), 0.01);
   }

   @Test
   public void logProbability() throws Exception {
      assertEquals(Math.log(0.205), binomial.logProbability(6), 0.01);
   }
}